requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blosc=1.20.1=he1b5a44_0
brotli=1.0.9=he1b5a44_0
brunsli=0.1=he1b5a44_0
bzip2=1.0.8=h516909a_3
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py37hc8dfbb8_0
charls=2.1.0=he1b5a44_2
cloudpickle=1.6.0=py_0
cycler=0.10.0=py_2
cytoolz=0.10.1=py37h516909a_0
dask-core=2.26.0=py_0
decorator=4.4.2=py_0
freetype=2.10.2=he06d7ca_0
giflib=5.2.1=h516909a_2
imagecodecs=2020.5.30=py37hda6ee5b_2
imageio=2.9.0=py_0
ipython=5.8.0=py37_1
ipython_genutils=0.2.0=py_1
joblib=0.16.0=py_0
jpeg=9d=h516909a_0
jxrlib=1.1=h516909a_2
kiwisolver=1.2.0=py37h99015e2_0
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.35=h769bd43_9
lerc=2.2=he1b5a44_0
libaec=1.0.4=he1b5a44_1
libblas=3.8.0=17_openblas
libcblas=3.8.0=17_openblas
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h24d8f2e_16
libgfortran-ng=7.5.0=hdf63c60_16
libgomp=9.3.0=h24d8f2e_16
liblapack=3.8.0=17_openblas
libopenblas=0.3.10=pthreads_hb3c22a3_4
libpng=1.6.37=hed695b0_2
libstdcxx-ng=9.3.0=hdf63c60_16
libtiff=4.1.0=hc7e4089_6
libwebp-base=1.1.0=h516909a_3
libzopfli=1.0.3=he1b5a44_0
lz4-c=1.9.2=he1b5a44_3
matplotlib-base=3.3.2=py37hd478181_0
ncurses=6.2=he1b5a44_1
networkx=2.5=py_0
numpy=1.19.1=py37h7ea13bd_2
olefile=0.46=py_0
openjpeg=2.3.1=h981e76c_3
openssl=1.1.1g=h516909a_1
pandas=1.1.2=py37h3340039_0
pexpect=4.8.0=py37hc8dfbb8_1
pickleshare=0.7.5=py37hc8dfbb8_1001
pillow=7.2.0=py37h718be6c_1
pip=20.2.3=py_0
prompt_toolkit=1.0.15=py_1
ptyprocess=0.6.0=py_1001
pygments=2.7.0=py_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.7.8=h6f2ec95_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.1=pyh9f0ad1d_0
pywavelets=1.1.1=py37h161383b_2
pyyaml=5.3.1=py37h8f50634_0
readline=8.0=he28a2e2_2
scikit-image=0.17.2=py37h9fdb41a_2
scikit-learn=0.23.2=py37h6785257_0
scipy=1.5.2=py37hb14ef9d_0
setuptools=49.6.0=py37hc8dfbb8_0
shap=0.35.0=py37h0da4684_0
simplegeneric=0.8.1=py_1
six=1.15.0=pyh9f0ad1d_0
snappy=1.1.8=he1b5a44_3
sqlite=3.33.0=h4cf870e_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.9.3=py_0
tk=8.6.10=hed695b0_0
toolz=0.10.0=py_0
tornado=6.0.4=py37h8f50634_1
tqdm=4.49.0=pyh9f0ad1d_0
traitlets=5.0.4=py_0
wcwidth=0.2.5=pyh9f0ad1d_1
wheel=0.35.1=pyh9f0ad1d_0
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zfp=0.5.5=he1b5a44_1
zlib=1.2.11=h516909a_1009
zstd=1.4.5=h6597ccf_2
